+++ lib/flutter_carplay.dart
@@ -33,9 +33,13 @@ export 'package:flutter_carplay/models/information/information_template.dart'
     show CPInformationTemplate;
 export 'package:flutter_carplay/models/list/list_constants.dart'
     show CPListItemAccessoryTypes, CPListItemPlayingIndicatorLocations;
+export 'package:flutter_carplay/models/list/list_image_row_item.dart'
+    show CPListImageRowItem;
 export 'package:flutter_carplay/models/list/list_item.dart' show CPListItem;
 export 'package:flutter_carplay/models/list/list_section.dart'
     show CPListSection;
+export 'package:flutter_carplay/models/list/list_template_item.dart'
+    show CPListTemplateItem;
 export 'package:flutter_carplay/models/list/list_template.dart'
     show CPListTemplate;
 export 'package:flutter_carplay/models/poi/poi.dart' show CPPointOfInterest;