+++ lib/helpers/carplay_helper.dart
@@ -22,7 +22,7 @@ class FlutterCarplayHelper {
         for (var list in listTemplates) {
           for (var section in list.sections) {
             for (var item in section.items) {
-              if (item.uniqueId == elementId) {
+              if (item is CPListItem && item.uniqueId == elementId) {
                 listItem = item;
                 break l1;
               }
@@ -32,6 +32,39 @@ class FlutterCarplayHelper {
       }
     }
     return listItem;
+  }
+
+  CPListImageRowItem? findCPListImageRowItem({
+    required List<dynamic> templates,
+    required String elementId,
+  }) {
+    CPListImageRowItem? imageRowItem;
+    l1:
+    for (var t in templates) {
+      final List<CPListTemplate> listTemplates = [];
+      if (t is CPTabBarTemplate) {
+        for (var template in t.templates) {
+          if (template is CPListTemplate) {
+            listTemplates.add(template);
+          }
+        }
+      } else if (t is CPListTemplate) {
+        listTemplates.add(t);
+      }
+      if (listTemplates.isNotEmpty) {
+        for (var list in listTemplates) {
+          for (var section in list.sections) {
+            for (var item in section.items) {
+              if (item is CPListImageRowItem && item.uniqueId == elementId) {
+                imageRowItem = item;
+                break l1;
+              }
+            }
+          }
+        }
+      }
+    }
+    return imageRowItem;
   }
 
   String makeFCPChannelId({String? event = ''}) =>