+++ lib/models/list/list_item.dart
@@ -1,10 +1,11 @@
 import 'package:flutter_carplay/controllers/carplay_controller.dart';
 import 'package:flutter_carplay/helpers/enum_utils.dart';
 import 'package:flutter_carplay/models/list/list_constants.dart';
+import 'package:flutter_carplay/models/list/list_template_item.dart';
 import 'package:uuid/uuid.dart';
 
 /// A selectable list item object that appears in a list template.
-class CPListItem {
+class CPListItem implements CPListTemplateItem {
   /// Unique id of the object.
   final String _elementId = const Uuid().v4();
 
@@ -53,6 +54,7 @@ class CPListItem {
     this.accessoryType,
   });
 
+  @override
   Map<String, dynamic> toJson() => {
         '_elementId': _elementId,
         'text': text,
@@ -128,6 +130,7 @@ class CPListItem {
     FlutterCarPlayController.updateCPListItem(this);
   }
 
+  @override
   String get uniqueId {
     return _elementId;
   }